Directory restructuring in tests/Test/Workflow
